backend/requirements.txt
flask
flask-cors
sseclient
python-dotenv